refactor(ai-service): Steps 12 + 14 — consolidate config, delete analyzers/
Two cleanup steps bundled — both are pure deletes / moves with no
behavioural change.

Step 12 — single source of truth for configuration:
  app/core/__init__.py       →  app/infra/config/__init__.py
                                  (re-exports `settings`)
  app/core/config.py         →  app/infra/config/settings.py
  app/config/cache_config.py →  app/infra/config/cache.py
  app/config/pipeline_config.py → app/infra/config/pipeline.py
The two old packages (app/core/, app/config/) are deleted entirely.

Step 14 — drop the empty placeholder package:
  app/analyzers/__init__.py  →  deleted
  app/analyzers/             →  removed

Importers updated (12 files):
  • `from app.core import settings`           → `from app.infra.config import settings`
  • `from app.core.config import settings`    → `from app.infra.config.settings import settings`
  • `from app.config.pipeline_config import …` → `from app.infra.config.pipeline import …`
  • `from app.config.cache_config import …`   → `from app.infra.config.cache import …`

Validated:
  • `settings.app_name` resolves through the new path
  • app.core, app.config, app.analyzers all raise ImportError (gone)
  • /health 200

What's left under app/services/, app/models/, app/skill_manager/,
app/core/, app/config/, app/analyzers/: nothing. All eight pre-refactor
"junk drawer" / vestigial folders are deleted.

Step 13 (`POST /annotation/run` endpoint to decouple Streamlit's
in-process pipeline imports) is the only remaining item in the
diagram's migration map; tracking separately.
