--- acla_ai_service/app/analyzers/__init__.py
@@ -1,5 +1,0 @@
-"""
-Analyzers module initialization
-"""
-
-__all__ = []
