+++ acla_ai_service/app/api/health.py
@@ -6,7 +6,7 @@ from fastapi import APIRouter
 from datetime import datetime
 from app.api.schemas.query import HealthResponse
 from app.integrations.backend.client import backend_service
-from app.core import settings
+from app.infra.config import settings
 
 router = APIRouter(tags=["health"])
