--- acla_ai_service/app/core/__init__.py
@@ -1,7 +1,0 @@
-"""
-Core module initialization
-"""
-
-from .config import settings
-
-__all__ = ["settings"]
