+++ acla_ai_service/app/integrations/backend/client.py
@@ -14,7 +14,7 @@ import uuid
 import re
 import numpy as np
 from pathlib import Path
-from app.core import settings
+from app.infra.config import settings
 from app.integrations.backend.schemas import ActiveModelData
 logger = logging.getLogger(__name__)
