+++ acla_ai_service/app/ml/segment_classifier/service.py
@@ -476,7 +476,7 @@ class SegmentClassifierService:
 
     async def train_model(self, epochs=10, batch_size=32, learning_rate=0.001, val_split=0.1):
         """Train the LSTM Classifier using streaming data with train/val split."""
-        from app.config.pipeline_config import PipelineConfig
+        from app.infra.config.pipeline import PipelineConfig
         cache_key = PipelineConfig().annotation_cache_key
         
         train_key = f"{cache_key}_train"
@@ -980,7 +980,7 @@ class SegmentClassifierService:
             chunk_segments.append(segment.to_dict())
             
         # Cache segments
-        from app.config.pipeline_config import PipelineConfig
+        from app.infra.config.pipeline import PipelineConfig
         cache_key = PipelineConfig().segments_cache_key
         
         if chunk_segments:
